+++ src/mc_classifier_pipeline/label_studio_uploader.py
@@ -53,7 +53,7 @@ def parse_args():
         description="Upload formatted Label Studio tasks JSON to a specified Label Studio project via API.",
         formatter_class=argparse.RawDescriptionHelpFormatter,
         epilog="""
-            Examples:
+            Examples: 
             # Upload the default tasks file to Label Studio project with id 4
             python src/mc_classifier_pipeline/label_studio_uploader.py --project_id 4
             # Upload a custom tasks file to Label Studio project with id 100